fix: refactor + fix dos
